Move custom gltf code to dedicated lib
