+++ Cargo.toml
@@ -19,6 +19,8 @@ edition = "2021"
 
 [workspace.dependencies]
 clap = { version = "4.3.12", features = ["derive"] }
+gltf = { version = "=1.2.0", features = [ "extras", "import", "names" ] }
+gltf-json = { version = "=1.2.0", features = [ "names" ] }
 grim = { path = "core/grim" }
 itertools = "0.11.0"
 lazy_static = "1.4.0"
