+++ core/grim/Cargo.toml
@@ -10,8 +10,9 @@ bitstream-io = { version = "1.7.0", optional = true }
 ffmpeg-next = { version = "6.0.0", optional = true }
 flate2 = "1.0.26"
 fon = { version = "0.6.0", optional = true }
-gltf = { version = "1.2.0", optional = true, features = [ "extras", "import", "names" ] }
-gltf-json = { version = "1.2.0", optional = true, features = [ "names" ] }
+gltf = { workspace = true, optional = true }
+gltf-json = { workspace = true, optional = true }
+grim_gltf = { path = "../grim_gltf", optional = true }
 grim_macros = { path = "../grim_macros" }
 grim_traits = { path = "../grim_traits" }
 half = { version = "2.3.1", default-features = false }
@@ -38,7 +39,7 @@ rstest = "0.18.1"
 audio = [ "bitstream-io", "fon", "wav" ]
 audio_experimental = [ "ffmpeg-next" ]
 midi = [ "midly" ]
-model = [ "gltf", "gltf-json", "nalgebra", "serde" ]
+model = [ "gltf", "gltf-json", "grim_gltf", "nalgebra", "serde" ]
 python = [ "pyo3" ]
 
 [[bench]]
