+++ core/grim_gltf/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "grim_gltf"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+
+[dependencies]
+gltf = { workspace = true }
+gltf-json = { workspace = true }
+itertools = { workspace = true }
+serde = { workspace = true }
